refactor: move browser room facade to owner
